refactor: Move temporary file to new module
